pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tikbot"
version = "0.1.0"
description = "TikTok Live Bot with Serena AI Integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "TikTokLive>=3.0.0",
    "pyee>=8.2.0,<12.0.0",
    "websockets>=12.0",
    "aiofiles",
    "pydantic>=2.0.0",
    "pyyaml",
    "python-dotenv",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2",
    "requests",
    "colorlog",
    "pyttsx3>=2.90",
    "gtts>=2.0.0",
    "pygame>=2.1.0",
]

[project.optional-dependencies]
tts = [
    "pyttsx3",
    "gTTS",
    "azure-cognitiveservices-speech",
]
overlay = [
    "pillow>=10.0.0",
    "opencv-python",
]

[project.scripts]
tikbot = "tikbot.main:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/tikbot"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/pyproject.toml",
]